refactor to `useOnUnmount` hook
This will allow us to make the code relatively similar between React and
Vue.
